refactor piepline code
